+++ src/main/java/roomescape/auth/AuthController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.constant.AuthConstant;
 import roomescape.member.MemberService;
 import roomescape.member.dto.MemberLoginCheckResponse;
 import roomescape.member.dto.MemberLoginRequest;
