+++ src/main/java/roomescape/auth/JwtTokenManager.java
@@ -5,6 +5,7 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import roomescape.constant.AuthConstant;
 import roomescape.member.Member;
 
 @Component
