+++ src/main/java/roomescape/config/WebConfig.java
@@ -1,12 +1,12 @@
-package roomescape.common;
+package roomescape.config;
 
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.auth.AdminHandlerInterceptor;
-import roomescape.auth.LoginMemberArgumentResolver;
+import roomescape.global.AdminHandlerInterceptor;
+import roomescape.global.LoginMemberArgumentResolver;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
