+++ src/main/java/roomescape/global/CookieUtil.java
@@ -1,7 +1,8 @@
-package roomescape.auth;
+package roomescape.global;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
+import roomescape.constant.AuthConstant;
 
 public class CookieUtil {
