+++ src/main/java/roomescape/global/LoginMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.global;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
@@ -7,6 +7,9 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import roomescape.auth.JwtTokenManager;
+import roomescape.constant.AuthConstant;
+import roomescape.global.CookieUtil;
 import roomescape.member.dto.LoginMember;
 
 @Component
